requirements.txt
numpy==1.22.3
requests==2.27.1
Pillow==9.0.1